pyproject.toml
[tool.poetry]
name = "neurobot"
version = "0.1.0"
description = ""
authors = ["L1mple <L1mple@rambler.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.19.0"
pytest = "^7.2.0"
flake8 = "^5.0.4"
flake8-bugbear = "^22.4.25"
flake8-comprehensions = "^3.9.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.2.1"
pep8-naming = "^0.12.1"
flake8-string-format = "^0.3.0"
flake8-builtins = "^1.5.3"
flake8-black = "^0.3.3"
flake8-expression-complexity = "^0.0.11"
flake8-annotations-complexity = "^0.0.7"
flake8-requirements = "^1.5.3"
black = "^22.3.0"
flake8-isort = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
